Force schema update and enable plugin
Add functionality to allow an environment variable to indicate that the
COmanage database schema should be updated (./Console/cake database).
This is helpful when deploying a new version of the code that does not
result in a change in the version number and so upgradeVersion does not
fire. An example of this scenario is when new code is introduced in the
develop branch but before a release happens.

Also add functionality to allow an environment variable to indicate that
a non-core plugin should be enabled.