Set ownership on distribution LDAP files
Add logic to perform a chown openldap:openldap on the directory
files from the Debian distribution so that slapd always may start
even when the numeric uid has changed, for example when /etc/passwd
is mounted in. This was previously done for the directory files being
used (database with n 2) but not for those that are not being
used (database with n 1).